fix(security): merge dead cache-control middleware into active CSP middleware via sequence()
